Split storage volume APIs to libvirtApi/storageVolume.js
